AlquilerDto implementado junto a sus mappers

+++ src/main/java/com/bda/TrabajoPractico/repositorios/AlquilerRepository.java
@@ -3,5 +3,5 @@ package com.bda.TrabajoPractico.repositorios;
 import com.bda.TrabajoPractico.Entidades.Alquiler;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlquilerRepository extends JpaRepository<Alquiler,Integer> {
+public interface AlquilerRepository extends JpaRepository<Alquiler,Long> {
 }
